+++ NEWS.md
@@ -1,5 +1,7 @@
 # Next version
 
+# MESS 0.5.6
+
 *   Added the colCumSum function for fast computation of cumulative sums for each column in a matrix.
 *   Fixed a bug in auc 
 *   Fixed a bug in power_mcnemar_test and modified the default arguments [Thanks to John Lockwood]